+++ src/cli.rs
@@ -82,13 +82,13 @@ pub struct Opts {
 
 #[derive(Debug, StructOpt)]
 pub enum Command {
-    /// Update all addons and WeakAuras
+    /// Update all addons, WeakAura and Plater auras
     Update,
-    /// Update all addons from the command line then exit
+    /// Update all addons
     UpdateAddons,
-    /// Update all WeakAuras from the command line then exit
-    UpdateWeakauras,
-    /// Install an addon from the command line
+    /// Update all WeakAura and Plater auras
+    UpdateAuras,
+    /// Install an addon
     Install {
         #[structopt(parse(try_from_str = str_to_flavor), possible_values = &["retail","ptr","beta","classic","classic_ptr","classic_beta"])]
         /// flavor to install addon under
@@ -110,6 +110,14 @@ pub enum Command {
         destination: PathBuf,
         #[structopt(short, long, default_value = "zip", possible_values = &["zip", "zstd"])]
         compression_format: CompressionFormat,
+    },
+    /// Add a World of Warcraft path
+    PathAdd {
+        /// path to the World of Warcraft directory
+        path: PathBuf,
+        #[structopt(parse(try_from_str = str_to_flavor), possible_values = &["retail","ptr","beta","classic","classic_ptr","classic_beta"])]
+        /// flavor to use from the path. If none, we use all we find
+        flavor: Option<Flavor>,
     },
 }